requirements.txt
joblib==1.2.0
numpy==1.24.1
pandas==1.5.2
pygame==2.1.2
python-dateutil==2.8.2
pytz==2022.7
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sklearn-text-extensions==1.0.0
threadpoolctl==3.1.0
